feat: Drop stateCouchDB in drop_db.go
closes #466
